LPD-98705 Extract useTriggerFocusOnClose hook
